// main/dashboard/src/shared/types/voxlab/camera-types.ts
export interface CameraSettings {
    fov: number;
    near: number;
    far: number;
    positionX: number;
    positionY: number;
    positionZ: number;
    targetX: number;
    targetY: number;
    targetZ: number;
    dampingFactor: number;
    fitDistanceMultiplier: number;
    frontDistanceMultiplier: number;
}